minor cleanup. Fixed some leaks
git-svn-id: svn://orca.physics.unc.edu/Orca@10571 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/Halo/HaloDetectorView.m
@@ -80,7 +80,7 @@
  	[super drawRect:rect];
 }
 
-- (NSColor*) getColorForSet:(int)setIndex value:(float)aValue
+- (NSColor*) getColorForSet:(int)setIndex value:(unsigned long)aValue
 {
 	return [focalPlaneColorScale getColorForValue:aValue];
 }
